+++ CHANGELOG.md
@@ -12,6 +12,14 @@
 * Add Admin note field to the Holdings field mapping profile: View (UIDATIMP-1122)
 * Add Admin note field to the Item field mapping profile: View (UIDATIMP-1123)
 * Display Authority information on Data import log page: View (UIDATIMP-1053)
+* Fix and Cover ViewAllLogs component with unit tests: Test (UIDATIMP-1151)
+* Add UI validation for some Invoice field mapping profile fields (UIDATIMP-1134)
+* Add summary at top of individual import job's log (UIDATIMP-1153)
+* For the Data Import View all page, change from Load more to Paginated (UIDATIMP-1147)
+
+### Bugs fixed:
+* Data Import landing page log shows in old format instead of current format (UIDATIMP-1139)
+* Some issues with log searching in Juniper Bugfest/Smoke testing (UIDATIMP-1125)
 
 ## [5.1.2](https://github.com/folio-org/ui-data-import/tree/v5.1.2) (2022-03-24)